+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `--force` option for `start` command to send SIGKILL to instances
+
 ## [2.1.0] - 2020-07-17
 
 ### Added