+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `cartridge clean` command to remove instance(s) files
 - `--from` option for `create` command to use custom application templates
 
+### Changed
+
+- Update `cartridge` to 2.2.1
+
 ## [2.1.0] - 2020-07-17
 
 ### Added